Fixed TODO and a bug, and moved raycast to Map
Now you can not hit an active object if there is a node in front of it.
You can only select nodes if they are not farther than d. The raycasting
algorithm was sliced into raycast() and createHilightbox(). raycast()
was moved into Map so it can be used in hostile creatures' AIs.